+++ src/Pages/Cart/Cart.scss
@@ -36,124 +36,8 @@
     margin: 0 auto 200px auto;
     padding-top: 100px;
 
-    .freeDeliveryBar {
-      display: flex;
-      justify-content: space-around;
-      align-items: center;
-      position: relative;
-      width: 80%;
-      padding: 100px 0;
-      border-bottom: 1px solid grey;
-      font-size: 16px;
-      font-weight: 700;
-
-      .totalBar {
-        position: relative;
-        width: inherit;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-
-        .movingCharacter {
-          position: absolute;
-          left: 60px;
-          width: 70px;
-        }
-
-        .movingPrice {
-          position: absolute;
-          bottom: -70px;
-          left: 60px;
-          padding: 10px 15px;
-          border-radius: 5px;
-          background-color: grey;
-          color: $white;
-        }
-
-        .percentageBarLeft {
-          position: relative;
-          background-color: $yellow;
-          border-radius: 10px;
-          width: 5%;
-          height: 20px;
-        }
-
-        .percentageBarRight {
-          background-color: $lightGrey;
-          width: 80%;
-          height: 20px;
-          border-radius: 10px;
-        }
-      }
-    }
-
     .checkBox {
       width: 80%;
-
-      .checkBoxHeader {
-        display: flex;
-        align-items: center;
-        padding: 50px 0;
-
-        font-size: 16px;
-
-        .check {
-          margin-right: 10px;
-        }
-
-        .choiceNum {
-          margin-right: 30%;
-        }
-
-        .choiceOption {
-          margin-right: 15%;
-        }
-
-        .choiceQuantity {
-          margin-right: 15%;
-        }
-
-        .productTotal {
-          margin-right: 5%;
-        }
-
-        span {
-          font-size: 16px;
-        }
-      }
-    }
-
-    .totalPriceText {
-      display: flex;
-      align-items: center;
-      padding: 50px 0;
-      width: 80%;
-      border-top: 1px solid grey;
-      div {
-        padding: 0 9%;
-        font-size: 20px;
-      }
-    }
-
-    .totalPriceNum {
-      position: relative;
-      display: flex;
-      align-items: center;
-      padding: 50px 0;
-      width: 80%;
-      border-top: 1px solid $lightGrey;
-      border-bottom: 1px solid $lightGrey;
-
-      div {
-        padding: 0 9%;
-        font-size: 20px;
-        font-weight: 300;
-      }
-
-      .equal {
-        position: absolute;
-        left: 75%;
-      }
     }
 
     .orderButton {
