+++ src/Pages/Cart/Components/ProductInCart/ProductInCart.scss
@@ -1,4 +1,4 @@
-@import '../../../Styles/variables';
+@import '../../../../Styles/variables';
 
 .productInCart {
   .cartProductBox {
@@ -11,12 +11,12 @@
     img {
       width: 200px;
       margin: 20px;
+      border-radius: 10px;
     }
 
     .productNamePrice {
       display: flex;
       flex-direction: column;
-      width: 200px;
     }
 
     .productOption {
@@ -24,7 +24,6 @@
       left: 60%;
       display: flex;
       align-items: center;
-      width: 500px;
 
       .productQuantity {
         .quantityGroup {
@@ -50,7 +49,6 @@
       }
 
       .fa-trash {
-        margin-left: 20px;
         color: $lightGrey;
         font-size: 30px;
         margin-left: 70px;
